requirements.txt
aiogram==3.0.0
Jinja2==3.1.3
python-dotenv==1.0.0
icecream==2.1.3
